Make telemetry opt-in, with env.MC_TELEMETRY set to 1 or true
