+++ util/telemetry/src/lib.rs
@@ -83,48 +83,8 @@ pub fn start_block_span<T: Tracer>(
     block_span_builder(tracer, span_name, block_index).start(tracer)
 }
 
-cfg_if::cfg_if! {
-    if #[cfg(feature = "jaeger")] {
-        use displaydoc::Display;
-        use opentelemetry::{trace::TraceError, KeyValue, sdk};
+#[cfg(feature = "jaeger")]
+mod jaeger;
 
-        #[derive(Debug, Display)]
-        pub enum Error {
-            /// Trace error: {0}
-            Trace(TraceError),
-
-            /// Get hostname error: {0}
-            GetHostname(std::io::Error),
-
-            /// Failed converting hostname to string
-            HostnameToString,
-        }
-
-        pub fn setup_default_tracer_with_tags(service_name: &str, extra_tags: &[(&'static str, String)]) -> Result<sdk::trace::Tracer, Error> {
-            let local_hostname = hostname::get().map_err(Error::GetHostname)?;
-
-            let mut tags = vec![KeyValue::new(
-                "hostname",
-                local_hostname
-                    .to_str()
-                    .ok_or(Error::HostnameToString)?
-                    .to_owned(),
-            )];
-            for (key, value) in extra_tags.iter() {
-                tags.push(KeyValue::new(*key, value.clone()));
-            }
-
-            opentelemetry_jaeger::new_pipeline()
-                .with_service_name(service_name)
-                .with_trace_config(
-                    sdk::trace::Config::default()
-                        .with_resource(sdk::Resource::new(tags)))
-                .install_simple()
-                .map_err(Error::Trace)
-        }
-
-        pub fn setup_default_tracer(service_name: &str) -> Result<sdk::trace::Tracer, Error> {
-            setup_default_tracer_with_tags(service_name, &[])
-        }
-    }
-}
+#[cfg(feature = "jaeger")]
+pub use jaeger::{setup_default_tracer, setup_default_tracer_with_tags, Error};
